Bug 64465 - publish jmeter-bom and jmeter-dependencies-bom
ApacheJMeter_bom lists versions for JMeter artifacts only:
  ApacheJMeter_core, ApacheJMeter_components, ...
ApacheJMeter_dependencies-bom lists versions for third-party dependencies only
